Fixed.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@31433 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/beos/io-access.c
@@ -111,7 +111,7 @@ void io_access_store_long(WORD addr, DWORD value)
 #endif
 }
 
-BYTE io_access_read_long(WORD addr)
+DWORD io_access_read_long(WORD addr)
 {
 #ifdef __HAIKU__
     port_io_args args = { POKE_SIGNATURE, addr, 4, 0 };
